refactor: changed the local to implement client management functions
