+++ base-cli/cmd/static/auth/clients/list.go
@@ -2,7 +2,9 @@ package clients
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"net/http"
 
 	"github.com/magaluCloud/mgccli/beautiful"
 	"github.com/magaluCloud/mgccli/cmd/common/auth"
@@ -10,6 +12,34 @@ import (
 	"github.com/magaluCloud/mgccli/i18n"
 	"github.com/spf13/cobra"
 )
+
+type Clients struct {
+	UUID        string `json:"uuid,omitempty"`
+	ClientID    string `json:"client_id,omitempty"`
+	Name        string `json:"name,omitempty"`
+	Description string `json:"description,omitempty"`
+	Status      string `json:"status,omitempty"`
+	Scopes      []struct {
+		UUID string `json:"uuid"`
+		Name string `json:"name"`
+	} `json:"scopes,omitempty"`
+	ScopesDefault []struct {
+		UUID string `json:"uuid"`
+		Name string `json:"name"`
+	} `json:"scopes_default,omitempty"`
+	TermOfUse                        string   `json:"client_term_url,omitempty"`
+	ClientPrivacyTermUrl             string   `json:"client_privacy_term_url,omitempty"`
+	Audience                         []string `json:"audience,omitempty"`
+	OidcAudience                     []string `json:"oidc_audience,omitempty"`
+	AlwaysRequireLogin               bool     `json:"always_require_login,omitempty"`
+	BackchannelLogoutSessionRequired bool     `json:"backchannel_logout_session_required,omitempty"`
+	BackchannelLogoutUri             string   `json:"backchannel_logout_uri,omitempty"`
+	RefreshTokenCustomExpiresEnabled bool     `json:"refresh_token_custom_expires_enabled,omitempty"`
+	RefreshTokenExp                  int      `json:"refresh_token_exp,omitempty"`
+	AccessTokenExp                   int      `json:"access_token_exp,omitempty"`
+	RedirectURIs                     []string `json:"redirect_uris,omitempty"`
+	Icon                             string   `json:"icon,omitempty"`
+}
 
 type ListClientResult struct {
 	UUID                             string   `json:"uuid,omitempty"`
@@ -53,9 +83,7 @@ func ListCommand(ctx context.Context) *cobra.Command {
 
 // runList executa o processo de exibir todos clientes
 func runList(ctx context.Context, rawMode bool) error {
-	auth := ctx.Value(cmdutils.CTX_AUTH_KEY).(auth.Auth)
-
-	clients, err := auth.ListClients(ctx)
+	clients, err := listClients(ctx)
 	if err != nil {
 		return fmt.Errorf("erro ao listar os clientes: %w", err)
 	}
@@ -100,4 +128,52 @@ func runList(ctx context.Context, rawMode bool) error {
 	}
 
 	return nil
+}
+
+func listClients(ctx context.Context) ([]*Clients, error) {
+	authCtx := ctx.Value(cmdutils.CTX_AUTH_KEY).(auth.Auth)
+
+	config := authCtx.GetConfig()
+
+	client, err := auth.NewOAuthClient(config)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create OAuth client: %w", err)
+	}
+
+	httpClient := client.AuthenticatedHttpClientFromContext(ctx)
+	if httpClient == nil {
+		return nil, fmt.Errorf("programming error: unable to get HTTP Client from context")
+	}
+
+	r, err := http.NewRequestWithContext(
+		ctx,
+		http.MethodGet,
+		config.ClientsURLV2,
+		nil,
+	)
+	if err != nil {
+		return nil, err
+	}
+	r.Header.Set("Content-Type", "application/json")
+
+	resp, err := httpClient.Do(r)
+	if err != nil {
+		return nil, err
+	}
+
+	clients := []*Clients{}
+	if resp.StatusCode == http.StatusNoContent {
+		return clients, nil
+	}
+
+	if resp.StatusCode != http.StatusOK {
+		return nil, cmdutils.NewHttpErrorFromResponse(resp, r)
+	}
+
+	defer resp.Body.Close()
+	if err = json.NewDecoder(resp.Body).Decode(&clients); err != nil {
+		return nil, err
+	}
+
+	return clients, nil
 }
